参考/YCrispBiscuit-Tools/YCrispBiscuit-Tools/Server/YCrispBiscuit-Tools-Server/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
SQLAlchemy
python-multipart
pymysql
requests
Pillow